# API/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Citi+
  description: Documentation for Citi+ REST API.
  termsOfService: https://www.ratp.fr/gestion-de-vos-donnees-personnelles
  contact:
    email: l_laugier@etu-webschoolfactory.fr
  license:
    name: MIT License
    url: http://localhost:80/api/v1/documentation/LICENSE
  version: 1.0.0
...
servers:
- url: http://localhost:80/api/v1/
...
paths:
...
  /indices-qualite-air/{arrondissement}:
    get:
      tags:
      - airparif
      summary: Get air quality data by arrondissement number for the last 7 days
      description: Use "all" for an average of all arrondissements.
      operationId: DataGET
      parameters:
      - name: arrondissement
        in: path
        description: 'example: 15'
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/airQuality'
        500:
          description: Database error
          content: {}
...
  /indice-elevation/{arrondissement}:
    get:
      tags:
      - elevation
      summary: Get landscape elevation data by arrondissement number
      operationId: ArrondissementGET
      parameters:
      - name: arrondissement
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: number
                format: double
        500:
          description: Database error
          content: {}
  /maratp-clients:
    get:
      tags:
      - maratp
      summary: Get all clients
      operationId: ClientsGET
      parameters:
      - name: limit
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          type: string
      - name: genre
        in: query
        schema:
          type: string
          enum:
          - femme
          - homme
      - name: titre_transport
        in: query
        description: 'example: Navigo Mois'
        schema:
          type: string
      - name: frequence_transport
        in: query
        description: 'example: 1 fois par semaine'
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/client'
        500:
          description: Database error
          content: {}
    post:
      tags:
      - maratp
      summary: Create a new client
      operationId: ClientCreate
      requestBody:
        description: Client object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/client'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
        500:
          description: Database error
          content: {}
      x-codegen-request-body-name: body
  /maratp-client/{id}:
    get:
      tags:
      - maratp
      summary: Get a client by ID
      operationId: ClientGET
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
        500:
          description: Database error
          content: {}
    put:
      tags:
      - maratp
      summary: Update an existing client
      operationId: ClientUpdate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        description: Client object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/client'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
        500:
          description: Database error
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
      - maratp
      summary: Delete an existing client
      operationId: ClientDelete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/client'
        500:
          description: Database error
          content: {}
...
components:
  schemas:
    airQuality:
      type: object
      properties:
        O3:
          type: array
          items:
            type: number
            format: double
        NO2:
          type: array
          items:
            type: number
            format: double
        PM10:
          type: array
          items:
            type: number
            format: double
        PM25:
          type: array
          items:
            type: number
            format: double
...
    client:
      required:
      - identifiant
      type: object
      properties:
        identifiant:
          type: integer
          format: int64
        age:
          type: string
          example: "21"
        genre:
          type: string
          example: homme
        code_postal:
          type: string
          example: "93250"
        anciennete:
          type: integer
          format: int64
        abonne_alerting:
          type: boolean
          default: false
        alertes:
          type: string
          example: '[[''rer-b'', ''07:00:00'', ''09:00:00'', ''vendredi''], [''rer-b'',
            ''18:00:00'', ''20:00:00'', ''vendredi'']]'
        titre_transport:
          type: string
          example: Navigo Annuel
        frequence_transport:
          type: string
          example: Plusieurs fois par jour
        favoris_horaires:
          type: string
          example: '[[''poissy'', ''RER_A'', ''cergy_le_haut_poissy_saint_germain_en_laye''],
            [''poissy'', ''SNCF_52'', ''gare_saint_lazare'']]'
        favoris_adresses:
          type: string
          example: '[[''lieu de travail'', 94800]]'
...